+++ brightness/src/dpAux.h
@@ -1,0 +1,42 @@
+/*
+ *      dpAux.h                 (C) 2022-2023, Aurélien Croc (AP²C)
+ *      Modified in 2026 by Samsung-Galaxy-Book-12-Linux contributors:
+ *      error reporting and byte-safe interfaces.
+ *
+ *   This program is free software; you can redistribute it and/or modify it under
+ *   the terms of the GNU General Public License as published by the Free Software
+ *   Foundation; version 2 of the License.
+ *
+ *   This program is distributed in the hope that it will be useful, but WITHOUT
+ *   ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ *   FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+ *   details.
+ *
+ *   You should have received a copy of the GNU General Public License along with
+ *   this program; If not, see <http://www.gnu.org/licenses/>.
+ */
+#ifndef DPAUX_H
+#define DPAUX_H
+
+typedef int aux_t;
+
+
+typedef struct {
+    int             addr;
+    int             nVal;
+    unsigned char   val[3];
+} dpWrite_t;
+
+extern aux_t dpAuxOpen(const char *path);
+extern int dpAuxRead(aux_t aux, int addr);
+extern int dpAuxWrite(aux_t aux, int addr, unsigned char val);
+extern int dpAuxWrites(aux_t aux, dpWrite_t data[]);
+extern void dpAuxClose(aux_t aux);
+extern void dpAuxClearError(void);
+extern int dpAuxLastError(void);
+
+
+
+#endif /* DPAUX_H */
+
+/* vim: set expandtab tabstop=4 shiftwidth=4 smarttab tw=80 cin fenc=utf8 : */